refactor: Move servicescatalogcache package to libs
