+++ proxy/server/server.go
@@ -10,13 +10,13 @@ import (
 	"github.com/AdguardTeam/dnsproxy/proxy"
 	"github.com/getsentry/sentry-go"
 	"github.com/ivpn/dns/libs/logging"
+	"github.com/ivpn/dns/libs/servicescatalogcache"
 	"github.com/ivpn/dns/proxy/cache"
 	"github.com/ivpn/dns/proxy/cache/memory"
 	"github.com/ivpn/dns/proxy/collector/channel"
 	"github.com/ivpn/dns/proxy/config"
 	"github.com/ivpn/dns/proxy/filter"
 	"github.com/ivpn/dns/proxy/internal/asnlookup"
-	"github.com/ivpn/dns/proxy/internal/servicescatalogcache"
 	"github.com/ivpn/dns/proxy/model"
 	"github.com/ivpn/dns/proxy/requestcontext"
 	"github.com/miekg/dns"
